+++ src/niquests/sessions.py
@@ -222,6 +222,7 @@ class Session:
         "_keepalive_idle_window",
         "base_url",
         "quic_cache_layer",
+        "timeout",
     ]
 
     def __init__(
@@ -243,6 +244,7 @@ class Session:
         keepalive_delay: float | int | None = 3600.0,
         keepalive_idle_window: float | int | None = 60.0,
         base_url: str | None = None,
+        timeout: TimeoutType | None = None,
     ):
         """
         :param resolver: Specify a DNS resolver that should be used within this Session.
@@ -266,6 +268,7 @@ class Session:
         :param keepalive_idle_window: Delay expressed in seconds, in which we should send a PING frame after the connection
             being completely idle. This only applies to HTTP/2 onward.
         :param base_url: Automatically set a URL prefix (or base url) on every request emitted if applicable.
+        :param timeout: Default timeout configuration to be used if no timeout is provided in exposed methods.
         """
         if [disable_ipv4, disable_ipv6].count(True) == 2:
             raise RuntimeError("Cannot disable both IPv4 and IPv6")
@@ -313,6 +316,9 @@ class Session:
         self.resolver = create_resolver(resolver)
         #: Internal use, know whether we should/can close it on session close.
         self._own_resolver: bool = resolver != self.resolver
+
+        #: global timeout configuration
+        self.timeout = timeout
 
         #: Bind to address/network adapter
         self.source_address = source_address
@@ -472,7 +478,7 @@ class Session:
         cookies: CookiesType | None = None,
         files: MultiPartFilesType | MultiPartFilesAltType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = WRITE_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -555,10 +561,16 @@ class Session:
 
         # Send the request.
         send_kwargs = {
-            "timeout": timeout,
+            "timeout": timeout or self.timeout,
             "allow_redirects": allow_redirects,
         }
         send_kwargs.update(settings)
+
+        if send_kwargs["timeout"] is None:
+            send_kwargs["timeout"] = (
+                WRITE_DEFAULT_TIMEOUT if method in {"POST", "PUT", "DELETE", "PATCH"} else READ_DEFAULT_TIMEOUT
+            )
+
         resp = self.send(prep, **send_kwargs)
 
         return resp
@@ -571,7 +583,7 @@ class Session:
         headers: HeadersType | None = None,
         cookies: CookiesType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = READ_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -640,7 +652,7 @@ class Session:
         headers: HeadersType | None = None,
         cookies: CookiesType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = READ_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -709,7 +721,7 @@ class Session:
         headers: HeadersType | None = None,
         cookies: CookiesType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = READ_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -781,7 +793,7 @@ class Session:
         cookies: CookiesType | None = None,
         files: MultiPartFilesType | MultiPartFilesAltType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = WRITE_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -860,7 +872,7 @@ class Session:
         cookies: CookiesType | None = None,
         files: MultiPartFilesType | MultiPartFilesAltType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = WRITE_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -939,7 +951,7 @@ class Session:
         cookies: CookiesType | None = None,
         files: MultiPartFilesType | MultiPartFilesAltType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = WRITE_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,
@@ -1015,7 +1027,7 @@ class Session:
         headers: HeadersType | None = None,
         cookies: CookiesType | None = None,
         auth: HttpAuthenticationType | None = None,
-        timeout: TimeoutType | None = WRITE_DEFAULT_TIMEOUT,
+        timeout: TimeoutType | None = None,
         allow_redirects: bool = True,
         proxies: ProxyType | None = None,
         hooks: HookType[PreparedRequest | Response] | None = None,